Adding algo flow in new file. Testing timing
