+++ buy-sell-algorithm/predictions/train.py
@@ -226,184 +226,5 @@ class Train:
         else:
             print("Training on this data not implemented yet")
             return None
-        
-if __name__ == "__main__":
-    """
-        This setup tries to show how the trainer should be used.
-
-        The flow: 
-            - At tick = 0, get historical data. If this is first call to trainer, set predictions to most recent history. If not, set
-            predictions to already prepared next predictions 
-            - At every 15th tick, fill up next predictions buffers using a batch of 15 values stored in the live datas buffers
-            - At every other tick, do something else, including filling up data buffers
-
-        This assumes that the server tick is also at 0, here: `cycles = sum([[i for i in range(60)] for _ in range(3)], [])`. When the trainer gets 
-        called for the first time. So it should be the case that  if the live tick isn't at 0 yet, just uses the historical data as the prediction 
-        for the next cycle.
-
-        Just so I can plot nice graphs without having to wait for 5 minutes for each cycle, and show the training speed, I've commented out the
-        functions that get the live data, and used the most recent history to train the models for next predictions here: 
-        `next_predictions[data_name] += trainer.query_model(data_name, x, y, serve.parsed_data[data_name][x:y])`
-
-        In reality, we want to use the live buffers for training, and update them at the relevant time steps
-    """
-
-    trainer = Train(elitism=0.2, mutation_prob=0.08, mutation_power=0.1, max_epochs=20, num_of_histories=5, 
-                data_batch_size=15, nn_batch_size=60, parsed_data=serve.parsed_data)
-    loop = asyncio.get_event_loop()
-    
-    def new_cycle(data_buffers, predictions, next_predictions):
-        """
-            - Set prediction of current cycle to be the most recent history if this is first call to trainer at tick 0
-            - If this is not first call to trainer at tick 0, then next predictions should become current predictions
-            - Empty data and next predictions buffers, add current values into data buffers to begin with
-        """
-        start = time.time()
-
-        # at start of new cycle, prepare predictions for current cycle, and set correct historical data
-        serve.set_historical_prices()
-        trainer.change_historical_data(serve.parsed_data)
-
-        if(trainer.first_call()):
-            print("First call, assume predictions for all data is most recent cycle")
-            
-            for data_name, _ in trainer.histories_buffer.items():
-                previous, most_recent = trainer.get_synthetic_data(data_name)
-                trainer.histories_buffer[data_name] = previous[1:] + [most_recent]
-                predictions[data_name] = most_recent
-            
-        else:
-            print("Current predictions are ready")
-            if any([len(n) == 0 for n in next_predictions.values()]):
-                predictions = trainer.histories_buffer
-            else:
-                predictions = next_predictions.copy()
-
-        # empty data and next prediction buffers and add the current live values
-        serve.live_data()
-        
-        for data_name in ['buy_price', 'sell_price', 'demand']:
-            data_buffers[data_name] = []
-            next_predictions[data_name] = []
-            data_buffers[data_name].append(serve.parsed_data[data_name])
-
-        for n, p in predictions.items():
-            plot_datas([p], "Prediction", n)
-
-        return time.time() - start
-    
-    def prepare_next(i, starting_i, data_buffers, next_predictions):
-        """
-            - Prepare predictions for the next cycle using the data buffers you have so far
-            - Also add value at current tick into data buffer
-        """
-        start = time.time()
-        # prepare next predictions for next batch each time we are at tick 15, 30, 45, 60
-        # start a new thread that does this batch while another thread continues on with the for loop
-        
-        serve.live_data()
-
-        for data_name in ['buy_price', 'sell_price', 'demand']:
-            data_buffers[data_name].append(serve.parsed_data[data_name])
-        
-        # if we started at new cycle, then get nexts
-        # if we didn't 
-        if(not ((starting_i % 15 == 0) and (trainer.first_call())) or starting_i < 30):
-            if(0 < i  - starting_i < 15):
-                dist = i - starting_i
-            else:
-                dist = 15
-
-            if(i == 59):
-                x, y = 60-dist, 60
-            else:
-                x, y = batch_up((i-dist, i), dist)[0]
-
-
-            for data_name in ['buy_price', 'sell_price', 'demand']:
-                # print(x, y, data_buffers[data_name][x:y+1])   
-
-                if(next_predictions[data_name] == [] and x != 0):
-                    next_predictions[data_name] = trainer.histories_buffer[data_name][-1][:x]
-                    #print(trainer.histories_buffer[data_name], data_name)
-
-                next_predictions[data_name] += trainer.query_model(data_name, x, y, data_buffers[data_name][x:y+1])
-                #next_predictions[data_name] += trainer.query_model(data_name, x, y, serve.parsed_data[data_name][x:y])
-
-                print(len(next_predictions[data_name]))
-                assert(len(next_predictions[data_name]) % 15 == 0)
-
-                # print(x, y, time.time()-start)
-
-        return time.time()-start, next_predictions, data_buffers
-
-    def something_else(data_buffers, predictions):
-        start = time.time()
-        # do something else, must include filling data buffers
-
-        if(trainer.first_call()):
-            serve.set_historical_prices()
-            trainer.change_historical_data(serve.parsed_data)
-            
-            for data_name in ['buy_price', 'sell_price', 'demand']:
-                previous, most_recent = trainer.get_synthetic_data(data_name)
-                trainer.histories_buffer[data_name] = previous[1:] + [most_recent]
-                predictions[data_name] = most_recent
-
-        serve.live_data()
-        
-        print("doing some other stuff")
-
-        for data_name in ['buy_price', 'sell_price', 'demand']:
-            data_buffers[data_name].append(serve.parsed_data[data_name])
-
-        return time.time() - start
-        
-    def main():
-        data_buffers = {'buy_price':[], 'sell_price':[], 'demand':[]}
-        predictions = {'buy_price':[], 'sell_price':[], 'demand':[]}
-        next_predictions = {'buy_price':[], 'sell_price':[], 'demand':[]}
-
-        cycle_count = 0
-        
-        starting_i = serve.starting_tick()
-        i = starting_i
-
-        if(starting_i != 0):
-            for k, v in data_buffers.items():
-                data_buffers[k] = [0] * starting_i
-
-        print("Started at tick ", starting_i)
-
-        while True:
-            if(i == 0):
-                cycle_count += 1
-
-                print("Cycle ", cycle_count)
-                print()
-
-                time_taken = new_cycle(data_buffers, predictions, next_predictions)
-
-            elif((i % 15) == 0 or (i == 59)):
-                time_taken, next_predictions, data_buffers = prepare_next(i, starting_i, data_buffers, next_predictions)
-
-                print("Preparation took ", time_taken)
-            
-            else:
-                # trainer has not been called yet at this point => server didn't start at 0
-                # need to set predictions to be entire previous cycle
-                time_taken = something_else(data_buffers, predictions)
-                print("Something else took ", time_taken)
- 
-            if(5-time_taken < 0):
-                print("Something took too much time ", time_taken)
-                sys.exit(1)
-            else:
-                time.sleep(5-time_taken)
-                i = (i + 1) % 60           
-            print("Current tick ", i)
-
-    main()
-
-
+
